refactor: 런타임 목데이터 의존성 제거 및 설정 파일 분리
