+++ components/staff/finance-management/FinanceRequestListPage.tsx
@@ -8,9 +8,9 @@ import { useQuery } from "@tanstack/react-query";
 import { getPurchaseRequests } from "@/api/request/request.api";
 import type { PurchaseRequestStatus } from "@/api/request/request.dto";
 import ListPanel, { type ListPanelRow } from "@/components/staff/common/ListPanel";
+import { FINANCE_REQUESTS_PER_PAGE } from "@/components/staff/finance-management/financeRequestConstants";
 import StaffSidebar from "@/components/staff/common/StaffSidebar";
 import { useAuthSession } from "@/hooks/useAuthSession";
-import { FINANCE_REQUESTS_PER_PAGE } from "@/mocks/staffFinance";
 import { queryKeys } from "@/lib/queryKeys";
 import { colors, layout, radii, spacing, typography } from "@/styles/tokens";
 import { formatUtcToKstShortDate } from "@/utils/formatUtcToKstShortDate";
